requirements.txt
# Core application dependencies for the Agora framework.
# Use `pip install -r requirements.txt` to install.

google-adk==0.2.0
google-generativeai==0.7.2
pydantic==2.8.2
python-dotenv==1.0.1

litellm==1.41.14